# internal/model/dto/checkout.go
package dto

import "github.com/gofrs/uuid/v5"

type CheckoutRequest struct {
	Items []CheckoutItem `json:"items"`
}

type CheckoutItem struct {
	ProductID uuid.UUID `json:"product_id"`
	Quantity  int       `json:"quantity"`
}